refactor: update folder names
