+++ dumps/note.md
@@ -1,18 +1,18 @@
 Special cases should be handled specially:
 
-- biblegateway:
+- biblegateway.com:
   - 1cor-1
   - 2sam-23
   - gen-2
   - ps-36
   - matt-6
-- ktcgkpv:
+- ktcgkpv.org:
   - 1 cr-1
   - 2 sm-23
   - st-2
   - đnl-19
   - nkm-7
-- bibledotcom:
+- bible.com:
   - 2sa-23
   - gen-2
   - mat-6
