+++ src/bible.com/get-paragraph.ts
