+++ src/bible.com/get-version.ts
