+++ src/biblegateway.com/get-version.ts
