+++ src/biblegateway.com/main.ts
@@ -2,29 +2,40 @@
 import path from 'path';
 import { type Prisma } from '@prisma/client';
 import { Agent, setGlobalDispatcher } from 'undici';
-import { getAll } from '@/ktcgkpv/get-all';
-import { getBook } from '@/ktcgkpv/get-book';
-import { versionMapping } from '@/ktcgkpv/mapping';
+import { getAll } from '@/biblegateway.com/get-all';
+import { getBook } from '@/biblegateway.com/get-book';
+import { getPsalmMeta } from '@/biblegateway.com/get-psalm-meta';
+import { getVersion } from '@/biblegateway.com/get-version';
 import { withCheckpoint } from '@/lib/checkpoint';
 import prisma from '@/prisma/prisma';
 
 setGlobalDispatcher(new Agent({ connect: { timeout: 60_000 } }));
 
 (async () => {
-  const versionCode = 'KT2011' satisfies keyof typeof versionMapping;
-
-  await getBook(versionCode);
+  await getVersion();
 
   const version = await prisma.version.findFirstOrThrow({
     where: {
-      code: versionCode,
+      code: 'BD2011',
       language: {
-        webOrigin: 'https://ktcgkpv.org/',
+        webOrigin: 'https://www.biblegateway.com',
       },
     },
     include: {
       formats: true,
     },
+  });
+
+  const versionFormat = await prisma.versionFormat.findFirstOrThrow({
+    where: {
+      versionId: version.id,
+      type: 'ebook',
+    },
+  });
+
+  await getBook({
+    type: versionFormat.type,
+    ref: versionFormat.ref,
   });
 
   const books = await prisma.book.findMany({
@@ -64,14 +75,17 @@ setGlobalDispatcher(new Agent({ connect: { timeout: 60_000 } }));
       filePath: path.join(
         __dirname,
         '../../../dist',
-        `ktcgkpv.org-${version.code}-checkpoint.json`,
+        `bible.com-${version.code}-checkpoint.json`,
       ),
     });
 
   for await (const checkpoint of chapterCheckpoint) {
     const chapter = checkpoint.params;
 
-    await getAll(chapter, versionCode);
+    await getAll(chapter);
+    if (chapter.book.code === 'ps') {
+      await getPsalmMeta(chapter);
+    }
 
     setCheckpointComplete(checkpoint.id, true);
   }
