+++ src/generate-bible-data/main.ts
@@ -4,5 +4,5 @@ import {
 } from '@/generate-bible-data/generate-data';
 
 (async () => {
-  await generateBibleData(presets.ktcgkpv, 'http://localhost:8081/api');
+  await generateBibleData(presets['ktcgkpv.org'], 'http://localhost:8081/api');
 })();
