+++ src/ktcgkpv.org/get-all-proper-name.ts
